+++ bin/updateSheetByApi.sh
@@ -3,6 +3,7 @@
 set -e
 
 dist/osx-x64/GuildWarsConsole gw:bank
+dist/osx-x64/GuildWarsConsole gw:char
 
 dist/osx-x64/GuildWarsConsole gw:check \
     configuration/.check-pvp-armor.json \