+++ src/Command/BankCommand/BankCommand.fs
@@ -29,54 +29,8 @@ module Bank =
 
             output.Section "Fetching item details"
             let! pricedItems =
-                let chunks =
-                    bank
-                    |> List.filter (fun { Count = count } -> count > 0)
-                    |> tee (List.length >> sprintf "<c:dark-yellow>Note:</c> There are <c:magenta>%A</c> items after filter out uncountable items" >> output.Message >> output.NewLine)
-                    |> List.chunkBySize 100
-
-                let progress =
-                    chunks
-                    |> List.length
-                    |> output.ProgressStart (chunks |> List.length |> sprintf "Fetching details for %A chunks ...")
-
-                chunks
-                |> List.map (fun items -> asyncResult {
-                    let itemIds =
-                        items
-                        |> List.map (fun { Id = id } -> id)
-
-                    let! prices =
-                        itemIds
-                        |> GuildWars.fetchItemPrices
-
-                    let! itemsInfo =
-                        itemIds
-                        |> GuildWars.fetchItems
-
-                    return
-                        items
-                        |> List.map (fun item ->
-                            {
-                                ItemInfo = itemsInfo.[item.Id]
-                                InventoryItem = item
-                                Price =
-                                    prices
-                                    |> Map.tryFind item.Id
-                                    |> function
-                                        | Some float -> float
-                                        | _ -> 0.0
-                            }
-                        )
-                        |> List.filter (fun i -> i.Price > 0.0)
-                        |> tee (fun _ -> progress |> output.ProgressAdvance)
-                })
-                |> AsyncResult.ofSequentialAsyncResults (sprintf "Error while fetching bank item details: %A")
-                <!> List.concat
-                >>* (fun items ->
-                    progress |> output.ProgressFinish
-                    items |> List.length |> sprintf "Fetched %A items" |> output.Success
-                )
+                bank
+                |> GuildWars.fetchPricedItems output GuildWars.BulkMode.Single
 
             pricedItems
             |> List.sortByDescending (fun item -> item.Price)